Refactor unit tests structure
